Edit SingleDigit.java

+++ src/tamrin2/SingleDigit.java
@@ -6,13 +6,13 @@ public class SingleDigit {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.print("Enter n : ");
-        int n = input.nextInt();
-        int sum = 0;
+        long n = input.nextLong();
+        long sum = 0;
         do {
             sum = 0;
-            int length = String.valueOf(n).length();
-            for (int i = 0; i < length; i++) {
-                char figure = String.valueOf(n).charAt(i);
+            long length = String.valueOf(n).length();
+            for (long i = 0; i < length; i++) {
+                char figure = String.valueOf(n).charAt((int) i);
                 sum += Integer.parseInt(String.valueOf(figure));
             }
             if (sum >= 10)
